Rename scrollable_sheet_extent to scrollable_sheet_position
