+++ lib/src/scrollable/sheet_scrollable.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
 import '../foundation/sheet_position_scope.dart';
-import 'scrollable_sheet_extent.dart';
+import 'scrollable_sheet_position.dart';
 import 'sheet_content_scroll_position.dart';
 
 class SheetScrollable extends StatefulWidget {
